+++ GreasyPopcornWebServicesServer/src/main/java/xyz/bcdi/greasypopcorn/webservices/UserResource.java
@@ -1,6 +1,7 @@
 package xyz.bcdi.greasypopcorn.webservices;
 
 import java.util.*;
+import java.util.stream.*;
 
 import javax.annotation.security.PermitAll;
 import javax.annotation.security.RolesAllowed;
@@ -9,7 +10,10 @@ import javax.ws.rs.core.*;
 import javax.ws.rs.core.Response.Status;
 
 import xyz.bcdi.greasypopcorn.core.User;
+import xyz.bcdi.greasypopcorn.core.User.UserBuilder;
 import xyz.bcdi.greasypopcorn.dbaccess.UserDAO;
+import xyz.bcdi.greasypopcorn.dbaccess.AbstractDatabaseAccessObject.SqlOperationEffect;
+import xyz.bcdi.greasypopcorn.webservices.security.SecurityUtils;
 
 @Path("users")
 public class UserResource extends AbstractResource {
@@ -18,7 +22,8 @@ public class UserResource extends AbstractResource {
 	@Produces(MediaType.APPLICATION_JSON)
 	@PermitAll
 	public List<User> getUsers() {
-		return UserDAO.getInstance().getUsers();
+		return UserDAO.getInstance().getUsers().stream()
+					.map(u -> UserBuilder.copyOf(u).withPassword(null).build()).collect(Collectors.toList());
 	}
 	
 	@GET
@@ -26,11 +31,12 @@ public class UserResource extends AbstractResource {
 	@Produces(MediaType.APPLICATION_JSON)
 	@PermitAll
 	public User getUser(@PathParam("username") String username) {
-		return UserDAO.getInstance().getUser(username);
+		User tmp = UserDAO.getInstance().getUser(username);
+		return UserBuilder.copyOf(tmp).withPassword(null).build();
 	}
-	
+
 	@PUT
-	@RolesAllowed("{Moderator, RegisteredUser}")
+	@RolesAllowed({"Moderator", "RegisteredUser"})
 	public Response replaceOrCreateUsers() {
 		return Response.status(Status.METHOD_NOT_ALLOWED).allow("GET", "POST").build();
 	}
@@ -38,22 +44,41 @@ public class UserResource extends AbstractResource {
 	@PUT
 	@Path("{username}")
 	@Consumes(MediaType.APPLICATION_JSON)
-	@RolesAllowed("{Moderator, RegisteredUser}")
-	public Response replaceOrCreateUser(User u, @PathParam("username") String username) {
-		return null;
+	@RolesAllowed({"Moderator", "RegisteredUser"})
+	public Response replaceOrCreateUser(User u, @PathParam("username") String username, @HeaderParam("Authorization") String authHeader) {
+		if (u.getUsername() == null || u.getUsername().isEmpty())
+			u = UserBuilder.copyOf(u).withUsername(username).build();
+		
+
+		if (!SecurityUtils.isUserAuthorized(authHeader, u))
+			return SecurityUtils.UNAUTHORIZED_RESPONSE.build();
+		
+		SqlOperationEffect opEffect = UserDAO.getInstance().replaceOrCreateUser(u);
+
+		if (opEffect == SqlOperationEffect.REPLACED)
+			return Response.status(Status.OK).build();
+		else if (opEffect == SqlOperationEffect.CREATED)
+			return Response.status(Status.CREATED).build();
+		else
+			return Response.status(Status.BAD_REQUEST).build();
 	}
 	
 	@POST
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
-	@RolesAllowed("Moderator")
-	public Response createUser(User u, @Context UriInfo uriInfo) {
-		return null;
+	@PermitAll
+	public Response createUser(User u, @Context UriInfo uriInfo, @HeaderParam("Authorization") String authHeader) {
+		/*if (!SecurityUtils.isUserAuthorized(authHeader, u))
+			return SecurityUtils.UNAUTHORIZED_RESPONSE.build();*/
+		
+		User result = UserDAO.getInstance().createUser(u);
+		String id = (result != null) ? result.getUsername() : null;
+		return buildResponseForCreateEntity(result, id, uriInfo);
 	}
 	
 	@POST
 	@Path("{username}")
-	@PermitAll
+	@RolesAllowed("Moderator")
 	public Response createUser(@PathParam("username") String username) {
 		return Response.status(Status.METHOD_NOT_ALLOWED).allow("GET", "PUT", "DELETE").build();
 	}
@@ -67,8 +92,12 @@ public class UserResource extends AbstractResource {
 	@DELETE
 	@Path("{username}")
 	@Produces(MediaType.APPLICATION_JSON)
-	@RolesAllowed("{Moderator, RegisteredUser}")
-	public Response deleteUser(@PathParam("username") String username) {
+	@RolesAllowed({"Moderator", "RegisteredUser"})
+	public Response deleteUser(@PathParam("username") String username, @HeaderParam("Authorization") String authHeader) {
+		User u = UserDAO.getInstance().getUser(username);
+		if (!SecurityUtils.isUserAuthorized(authHeader, u))
+			return SecurityUtils.UNAUTHORIZED_RESPONSE.build();
+		
 		boolean wasDeleted = UserDAO.getInstance().deleteUser(username);
 		return buildResponseForDeleteEntity(wasDeleted);
 	}
